ROFO-124 URL path를 좀더 좋게 변경
