+++ src/main/kotlin/kr/weit/roadyfoody/foodSpots/presentation/api/FoodSpotsController.kt
@@ -4,9 +4,7 @@ import jakarta.validation.Valid
 import jakarta.validation.constraints.Positive
 import jakarta.validation.constraints.Size
 import kr.weit.roadyfoody.auth.security.LoginUser
-import kr.weit.roadyfoody.common.dto.SliceResponse
 import kr.weit.roadyfoody.foodSpots.application.dto.FoodSpotsUpdateRequest
-import kr.weit.roadyfoody.foodSpots.application.dto.ReportHistoriesResponse
 import kr.weit.roadyfoody.foodSpots.application.dto.ReportRequest
 import kr.weit.roadyfoody.foodSpots.application.service.FoodSpotsCommandService
 import kr.weit.roadyfoody.foodSpots.application.service.FoodSpotsQueryService
@@ -15,13 +13,11 @@ import kr.weit.roadyfoody.foodSpots.validator.WebPImageList
 import kr.weit.roadyfoody.user.domain.User
 import org.springframework.http.HttpStatus.CREATED
 import org.springframework.http.MediaType.MULTIPART_FORM_DATA_VALUE
-import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PatchMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RequestPart
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
@@ -46,18 +42,6 @@ class FoodSpotsController(
         @RequestPart(required = false)
         reportPhotos: List<MultipartFile>?,
     ) = foodSpotsCommandService.createReport(user, reportRequest, reportPhotos)
-
-    @GetMapping("users/{userId}/histories")
-    override fun getReportHistories(
-        @PathVariable("userId")
-        userId: Long,
-        @Positive(message = "조회할 개수는 양수여야 합니다.")
-        @RequestParam(defaultValue = "10", required = false)
-        size: Int,
-        @Positive(message = "마지막 ID는 양수여야 합니다.")
-        @RequestParam(required = false)
-        lastId: Long?,
-    ): SliceResponse<ReportHistoriesResponse> = foodSpotsQueryService.getReportHistories(userId, size, lastId)
 
     @ResponseStatus(CREATED)
     @PatchMapping("/{foodSpotsId}")
