+++ src/main/kotlin/kr/weit/roadyfoody/user/application/service/UserQueryService.kt
@@ -1,16 +1,29 @@
 package kr.weit.roadyfoody.user.application.service
 
+import kr.weit.roadyfoody.common.dto.SliceResponse
+import kr.weit.roadyfoody.foodSpots.repository.FoodSpotsHistoryRepository
+import kr.weit.roadyfoody.foodSpots.repository.FoodSpotsPhotoRepository
+import kr.weit.roadyfoody.foodSpots.repository.ReportFoodCategoryRepository
+import kr.weit.roadyfoody.foodSpots.repository.getByHistoryId
+import kr.weit.roadyfoody.foodSpots.repository.getHistoriesByUser
 import kr.weit.roadyfoody.global.service.ImageService
 import kr.weit.roadyfoody.user.application.dto.UserInfoResponse
+import kr.weit.roadyfoody.user.application.dto.UserReportCategoryResponse
+import kr.weit.roadyfoody.user.application.dto.UserReportHistoriesResponse
+import kr.weit.roadyfoody.user.application.dto.UserReportPhotoResponse
 import kr.weit.roadyfoody.user.domain.User
 import kr.weit.roadyfoody.user.repository.UserRepository
 import kr.weit.roadyfoody.user.repository.getByUserId
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 
 @Service
 class UserQueryService(
     private val userRepository: UserRepository,
     private val imageService: ImageService,
+    private val foodSpotsHistoryRepository: FoodSpotsHistoryRepository,
+    private val foodSpotsPhotoRepository: FoodSpotsPhotoRepository,
+    private val reportFoodCategoryRepository: ReportFoodCategoryRepository,
 ) {
     fun getUserInfo(user: User): UserInfoResponse {
         val user = userRepository.getByUserId(user.id)
@@ -21,5 +34,27 @@ class UserQueryService(
             profileImageUrl,
             user.coin,
         )
+    }
+
+    @Transactional(readOnly = true)
+    fun getReportHistories(
+        userId: Long,
+        size: Int,
+        lastId: Long?,
+    ): SliceResponse<UserReportHistoriesResponse> {
+        val user = userRepository.getByUserId(userId)
+        val reportResponse =
+            foodSpotsHistoryRepository.getHistoriesByUser(user, size, lastId).map {
+                val reportPhotoResponse =
+                    foodSpotsPhotoRepository.getByHistoryId(it.id).map { photo ->
+                        UserReportPhotoResponse(photo, imageService.getDownloadUrl(photo.fileName))
+                    }
+                val reportCategoryResponse =
+                    reportFoodCategoryRepository.getByHistoryId(it.id).map { category ->
+                        UserReportCategoryResponse(category)
+                    }
+                UserReportHistoriesResponse(it, reportPhotoResponse, reportCategoryResponse)
+            }
+        return SliceResponse(reportResponse)
     }
 }
