+++ src/test/kotlin/kr/weit/roadyfoody/foodSpots/presentation/api/FoodSpotsControllerTest.kt
@@ -10,14 +10,10 @@ import io.mockk.just
 import io.mockk.mockk
 import io.mockk.runs
 import io.mockk.verify
-import kr.weit.roadyfoody.common.dto.SliceResponse
 import kr.weit.roadyfoody.foodSpots.application.service.FoodSpotsCommandService
 import kr.weit.roadyfoody.foodSpots.application.service.FoodSpotsQueryService
-import kr.weit.roadyfoody.foodSpots.fixture.TEST_FOOD_SPOTS_HAS_NEXT
-import kr.weit.roadyfoody.foodSpots.fixture.TEST_FOOD_SPOTS_LAST_ID
 import kr.weit.roadyfoody.foodSpots.fixture.TEST_FOOD_SPOTS_REQUEST_NAME
 import kr.weit.roadyfoody.foodSpots.fixture.TEST_FOOD_SPOTS_REQUEST_PHOTO
-import kr.weit.roadyfoody.foodSpots.fixture.TEST_FOOD_SPOTS_SIZE
 import kr.weit.roadyfoody.foodSpots.fixture.TEST_FOOD_SPOT_NAME_EMPTY
 import kr.weit.roadyfoody.foodSpots.fixture.TEST_FOOD_SPOT_NAME_INVALID_STR
 import kr.weit.roadyfoody.foodSpots.fixture.TEST_FOOD_SPOT_NAME_TOO_LONG
@@ -25,17 +21,14 @@ import kr.weit.roadyfoody.foodSpots.fixture.TEST_INVALID_TIME_FORMAT
 import kr.weit.roadyfoody.foodSpots.fixture.createMockPhotoList
 import kr.weit.roadyfoody.foodSpots.fixture.createOperationHoursRequest
 import kr.weit.roadyfoody.foodSpots.fixture.createTestFoodSpotsUpdateRequest
-import kr.weit.roadyfoody.foodSpots.fixture.createTestReportHistoriesResponse
 import kr.weit.roadyfoody.foodSpots.fixture.createTestReportRequest
 import kr.weit.roadyfoody.support.annotation.ControllerTest
 import kr.weit.roadyfoody.support.utils.ImageFormat
 import kr.weit.roadyfoody.support.utils.ImageFormat.WEBP
 import kr.weit.roadyfoody.support.utils.createMultipartFile
 import kr.weit.roadyfoody.support.utils.createTestImageFile
-import kr.weit.roadyfoody.support.utils.getWithAuth
 import kr.weit.roadyfoody.support.utils.multipartWithAuth
 import kr.weit.roadyfoody.support.utils.patchWithAuth
-import kr.weit.roadyfoody.user.fixture.TEST_USER_ID
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
 import org.springframework.mock.web.MockMultipartFile
 import org.springframework.test.web.servlet.MockMvc
@@ -283,59 +276,6 @@ class FoodSpotsControllerTest(
                                 )
                             }
                         }
-                    }
-                }
-            }
-
-            given("GET $requestPath/users/{userId}/histories Test") {
-                val response =
-                    SliceResponse(
-                        listOf(createTestReportHistoriesResponse()),
-                        TEST_FOOD_SPOTS_HAS_NEXT,
-                    )
-                every {
-                    foodSpotsQueryService.getReportHistories(any(), any(), any())
-                } returns response
-                `when`("정상적인 요청이 들어올 경우") {
-                    then("해당 유저의 리포트 이력을 반환한다.") {
-                        mockMvc
-                            .perform(
-                                getWithAuth("$requestPath/users/$TEST_USER_ID/histories")
-                                    .param("size", TEST_FOOD_SPOTS_SIZE.toString())
-                                    .param("lastId", TEST_FOOD_SPOTS_LAST_ID.toString()),
-                            ).andExpect(status().isOk)
-                    }
-                }
-
-                `when`("size와 lastId가 없는 경우") {
-                    every {
-                        foodSpotsQueryService.getReportHistories(any(), any(), any())
-                    } returns response
-                    then("기본값으로 해당 유저의 리포트 이력을 반환한다.") {
-                        mockMvc
-                            .perform(
-                                getWithAuth("$requestPath/users/$TEST_USER_ID/histories"),
-                            ).andExpect(status().isOk)
-                    }
-                }
-
-                `when`("size가 양수가 아닌 경우") {
-                    then("400을 반환") {
-                        mockMvc
-                            .perform(
-                                getWithAuth("$requestPath/users/$TEST_USER_ID/histories")
-                                    .param("size", "0"),
-                            ).andExpect(status().isBadRequest)
-                    }
-                }
-
-                `when`("lastId가 양수가 아닌 경우") {
-                    then("400을 반환") {
-                        mockMvc
-                            .perform(
-                                getWithAuth("$requestPath/users/$TEST_USER_ID/histories")
-                                    .param("lastId", "-1"),
-                            ).andExpect(status().isBadRequest)
                     }
                 }
             }
